refactor: Extract duplicate helper functions into shared utility module
Extract parsePattern, formatPatternName, getSeverityDisplay, and
groupOpportunitiesByPattern functions from detect.ts and track.ts into
src/utils/refactoring-utils.ts to eliminate code duplication.

Changes:
- Create src/utils/refactoring-utils.ts with shared refactoring utilities
- Remove duplicate functions from src/cli/refactor/detect.ts (lines 311-361)
- Remove duplicate formatPatternName from src/cli/refactor/track.ts
- Update analyze.ts to use shared getSeverityDisplaySimple, getRiskLevelDisplay, getPriorityDisplay, getSeverityIcon
- Update imports in all affected files

Benefits:
- Eliminates ~50 lines of duplicate code
- Centralizes refactoring display logic for consistency
- Reduces maintenance burden
- Improves code reusability
